requirements.txt
aiohttp==3.6.2
argon2-cffi==20.1.0
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.7.1
backcall==0.2.0
bleach==3.1.5
certifi==2020.12.5
cffi==1.14.1
chardet==3.0.4
colorama==0.4.3
cryptography==3.0
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
et-xmlfile==1.0.1
idna==2.10
importlib-metadata==4.3.1
ipykernel==5.3.4
ipython==7.17.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jdcal==1.4.1
jedi==0.17.2
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.6
jupyter-console==6.1.0
jupyter-core==4.6.3
jupyter-server-proxy==1.5.0
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.3.0
mistune==0.8.4
multidict==4.7.6
nbconvert==5.6.1
nbformat==5.0.7
notebook==6.1.2
numpy==1.19.1
opencv-python==4.5.1.48
openpyxl==3.0.5
packaging==20.4
pandas==1.1.0
pandocfilters==1.4.2
parso==0.7.1
pickleshare==0.7.5
Pillow==7.2.0
prometheus-client==0.8.0
prompt-toolkit==3.0.6
pycparser==2.20
pyglet==1.5.15
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
python-dotenv==0.18.0
pytube==10.6.0
pytz==2020.1
pywin32==228
pywinpty==0.5.7
pyzmq==19.0.2
qtconsole==4.7.5
QtPy==1.9.0
requests==2.25.1
scipy==1.5.2
Send2Trash==1.5.0
simpervisor==0.3
six==1.15.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
txaio==20.4.1
typing-extensions==3.7.4.3
urllib3==1.26.2
vpython==7.6.1
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
yarl==1.5.1
zipp==3.4.1
